requirements.txt
streamlit==1.37.0
langchain==0.2.1
langchain-core==0.2.3
langchain-community==0.2.1
langchain-google-genai==1.0.5
pinecone-client==4.1.1
langchain-pinecone==0.1.1
langchain-cohere==0.1.5
python-dotenv==1.0.1
tiktoken==0.7.0
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.4